refactor(server): extract shared lexical scanners into js-scan.js
